BAU - Replace curried Higher Order Function callbacks with EitherT
The current implementation uses HOF, in order to allow short circuiting
of the main block of Futures, but still provide a meaningful recovery type. This is
quite difficult to follow, as you have to trace the callbacks through multiple
functions.

These changes keep the same behaviour, but uses EitherT, so that the main logic
is inside of a for-comprehension, making the code more readable.

Have also replaced TraverseFuturesSequentially with foldLeftM - as no parallelism is
required here.